feat: migrate to vitest and merry in gardenfi/core
