--- packages/core/jest.config.ts
@@ -1,10 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'catalog',
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../coverage/catalog',
-};
