+++ packages/core/tsconfig.spec.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "outDir": "../dist/out-tsc",
     "module": "commonjs",
-    "types": ["jest", "node", "chrome"]
+    "types": ["node", "chrome"]
   },
   "include": [
     "jest.config.ts",
